# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "CheckoutSubscriptionCreate": {
        "description": "Start subscription checkout for a creator. Provide creator_id or creator_handle. success_url/cancel_url optional (from env).",
        "properties": {
          "cancel_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Url"
          },
          "creator_handle": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creator Handle"
          },
          "creator_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creator Id"
          },
          "success_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Url"
          }
        },
        "title": "CheckoutSubscriptionCreate",
        "type": "object"
      },
      "CheckoutSubscriptionOut": {
        "description": "Checkout URL for redirect.",
        "properties": {
          "checkout_url": {
            "title": "Checkout Url",
            "type": "string"
          }
        },
        "required": [
          "checkout_url"
        ],
        "title": "CheckoutSubscriptionOut",
        "type": "object"
      },
...
      "CreatorDiscoverPage": {
        "description": "Paginated list of discoverable creators.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CreatorDiscoverItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "CreatorDiscoverPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LedgerEntryCreate": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "currency": {
            "maxLength": 8,
            "minLength": 3,
            "title": "Currency",
            "type": "string"
          },
          "direction": {
            "pattern": "^(credit|debit)$",
            "title": "Direction",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          }
        },
        "required": [
          "account_id",
          "currency",
          "amount",
          "direction",
          "reference"
        ],
        "title": "LedgerEntryCreate",
        "type": "object"
      },
      "LedgerEntryOut": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account_id",
          "currency",
          "amount",
          "direction",
          "reference",
          "created_at"
        ],
        "title": "LedgerEntryOut",
        "type": "object"
      },
...
      "PostCreate": {
        "description": "Create post body.",
        "properties": {
          "asset_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Asset Ids",
            "type": "array"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "nsfw": {
            "default": false,
            "title": "Nsfw",
            "type": "boolean"
          },
          "type": {
            "pattern": "^(TEXT|IMAGE|VIDEO)$",
            "title": "Type",
            "type": "string"
          },
          "visibility": {
            "pattern": "^(PUBLIC|FOLLOWERS|SUBSCRIBERS)$",
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "type",
          "visibility"
        ],
        "title": "PostCreate",
        "type": "object"
      },
      "PostOut": {
        "description": "Post as returned by API (asset_ids only; client uses download endpoint).",
        "properties": {
          "asset_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Asset Ids",
            "type": "array"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "creator_user_id": {
            "format": "uuid",
            "title": "Creator User Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_locked": {
            "default": false,
            "description": "True when viewer cannot access content (teaser only).",
            "title": "Is Locked",
            "type": "boolean"
          },
          "locked_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When is_locked: SUBSCRIPTION_REQUIRED or FOLLOW_REQUIRED for UI copy.",
            "title": "Locked Reason"
          },
          "nsfw": {
            "title": "Nsfw",
            "type": "boolean"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "visibility": {
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "id",
          "creator_user_id",
          "type",
          "caption",
          "visibility",
          "nsfw",
          "created_at",
          "updated_at"
        ],
        "title": "PostOut",
        "type": "object"
      },
...
      "ProfileOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "created_at",
          "updated_at"
        ],
        "title": "ProfileOut",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserCreate": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "display_name"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserLogin": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin",
        "type": "object"
      },
      "UserOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfileOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "is_active",
          "profile",
          "created_at",
          "updated_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebhookAck": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "WebhookAck",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Zinovia Fans API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/signup": {
      "post": {
        "operationId": "auth_signup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signup",
        "tags": [
          "auth"
        ]
      }
    },
    "/billing/checkout/subscription": {
      "post": {
        "description": "Start Stripe Checkout for subscribing to a creator. Returns URL to redirect the fan. 501 if Stripe not configured.",
        "operationId": "billing_checkout_subscription",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutSubscriptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSubscriptionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Checkout Subscription",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/webhooks/stripe": {
      "post": {
        "description": "Stripe webhook: signature verified, idempotent by event id. Store event then process; set processed_at.",
        "operationId": "billing_webhooks_stripe",
        "parameters": [
          {
            "in": "header",
            "name": "Stripe-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stripe-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stripe Webhook",
        "tags": [
          "billing"
        ]
      }
    },
    "/creators": {
      "get": {
        "description": "List discoverable creators. Paginated. Optional search (handle/display_name ILIKE).",
        "operationId": "creators_list",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "Search by handle or display name",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by handle or display name",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorDiscoverPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Creators",
        "tags": [
          "creators"
        ]
      }
    },
...
    "/ledger/entries": {
      "post": {
        "description": "Admin-only. For internal tooling (e.g. balance adjustments). Not used by the web app.",
        "operationId": "ledger_create_entry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LedgerEntryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerEntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create ledger entry (admin)",
        "tags": [
          "ledger"
        ]
      }
    },
...
    "/posts": {
      "post": {
        "operationId": "posts_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "posts"
        ]
      }
    },
...
  }
}